Moves security tab into app folder
